+++ assets/javascripts/kitten/components/molecules/drag-and-drop-list/index.js
@@ -1,14 +1,34 @@
-import React, { useEffect, useRef, useState } from 'react'
+import React, { useState } from 'react'
+import {
+  DndContext,
+  closestCenter,
+  KeyboardSensor,
+  MouseSensor,
+  TouchSensor,
+  useSensor,
+  useSensors,
+  DragOverlay,
+} from '@dnd-kit/core'
+import {
+  arrayMove,
+  SortableContext,
+  sortableKeyboardCoordinates,
+  verticalListSortingStrategy,
+  useSortable,
+} from '@dnd-kit/sortable'
+import {
+  restrictToWindowEdges,
+  restrictToVerticalAxis,
+} from '@dnd-kit/modifiers'
+import { createPortal } from 'react-dom'
+
 import classNames from 'classnames'
 import PropTypes from 'prop-types'
-import DragonDrop from 'drag-on-drop'
 import { createGlobalStyle } from 'styled-components'
 
 import { pxToRem } from '../../../helpers/utils/typography'
 import { Button } from '../../../components/molecules/buttons/button'
 import { MenuIcon } from '../../../components/graphics/icons/menu-icon'
-
-export const BUTTON_SHIFT = pxToRem(40 + 15)
 
 const DragAndDropListStyles = createGlobalStyle`
   .k-DragAndDropList {
@@ -20,56 +40,64 @@ const DragAndDropListStyles = createGlobalStyle`
     margin: 0;
   }
 
+  .k-DragAndDropList__itemWrapper {
+    list-style: none;
+    display: block;
+    transform: translate3d(var(--translate-x, 0), var(--translate-y, 0), 0);
+    transform-origin: 0 0;
+    position: relative;
+    z-index: 1;
+
+    &:focus-within {
+      z-index: 2;
+    }
+  }
+
   .k-DragAndDropList__item {
     display: flex;
     gap: ${pxToRem(15)};
     align-items: center;
-    transition: opacity 0.2s ease;
 
-    &.gu-unselectable {
-      user-select: none !important;
+    &.k-DragAndDropList__item--isDragging:not(.k-DragAndDropList__item--dragOverlay) {
+      opacity: 0.4;
+
+      .k-DragAndDropList__item__button {
+        opacity: 0;
+      }
     }
 
-    &.gu-transit {
-      opacity: 0.2;
+    &.k-DragAndDropList__item--dragOverlay {
       cursor: move;
-    }
+      z-index: 150;
 
-    &.gu-mirror {
-      box-sizing: border-box;
-      position: fixed !important;
-      margin: 0 !important;
-      z-index: 9999 !important;
-      opacity: 0.9;
-    }
-
-    &.gu-hide {
-      display: none !important;
+      .k-DragAndDropList__item__button {
+        cursor: move;
+      }
     }
 
     .k-DragAndDropList__item__child {
-      flex: 1 0 100%;
-      max-width: 100%;
+      flex: 1 0 calc(100% - var(--dragAndDropList-button-shift, 0));
+      width: calc(100% - var(--dragAndDropList-button-shift, 0));
+      flex-grow: 1;
     }
 
-  }
-  .k-DragAndDropList__item--hasButton {
-    .k-DragAndDropList__item__button {
-      flex: 0 0 ${pxToRem(40)};
-      padding: 0;
-      width: ${pxToRem(40)};
-      box-sizing: border-box;
-      border-radius: ${pxToRem(20)};
-      cursor: grab;
-    }
+    &.k-DragAndDropList__item--hasButton {
+      .k-DragAndDropList__item__button {
+        flex: 0 0 ${pxToRem(40)};
+        padding: 0;
+        width: ${pxToRem(40)};
+        box-sizing: border-box;
+        border-radius: ${pxToRem(20)};
+        cursor: grab;
+      }
 
-    &.gu-mirror .k-DragAndDropList__item__button {
-      cursor: grabbing;
-    }
+      &.gu-mirror .k-DragAndDropList__item__button {
+        cursor: grabbing;
+      }
 
-    .k-DragAndDropList__item__child {
-      flex: 1 0 calc(100% - ${BUTTON_SHIFT});
-      max-width: calc(100% - ${BUTTON_SHIFT});
+      .k-DragAndDropList__item__child {
+        --dragAndDropList-button-shift: ${pxToRem(40 + 15)};
+      }
     }
   }
 `
@@ -80,113 +108,212 @@ export const DragAndDropList = ({
   onChange,
   a11yButtonLabel,
   a11yAnnouncement,
-  a11yButtonDescElement,
   a11yContainerLabelElement,
+  a11yInstructions,
   gap,
   style,
   showHandle,
   ...props
 }) => {
-  const listElement = useRef(null)
-  const [dragonInstance, setDragonInstance] = useState(null)
+  const [items, setItems] = useState(
+    React.Children.toArray(children).map(child => child.props.id),
+  )
+  const [childrenDict] = useState(
+    React.Children.toArray(children).reduce(
+      (acc, current) => ({ ...acc, [current.props.id]: current }),
+      {},
+    ),
+  )
+  const [activeId, setActiveId] = useState(null)
 
-  const dragonOptions = {
-    activeClass: 'k-DragAndDropList__item--active',
-    inactiveClass: 'k-DragAndDropList__item--inactive',
-    announcement: {
-      grabbed: element => a11yAnnouncement.grabbed(element.dataset.simpleName),
-      dropped: element => a11yAnnouncement.dropped(element.dataset.simpleName),
-      reorder: (element, items) => {
-        const position = items.indexOf(element) + 1
-        const name = element.dataset.simpleName
-        return a11yAnnouncement.reorder(name, position, items.length)
-      },
-      cancel: a11yAnnouncement.cancel,
+  const sensors = useSensors(
+    useSensor(MouseSensor, {
+      activationConstraint: { distance: 5 },
+    }),
+    useSensor(TouchSensor, {
+      activationConstraint: { delay: 250, tolerance: 5 },
+    }),
+    useSensor(KeyboardSensor, {
+      coordinateGetter: sortableKeyboardCoordinates,
+    }),
+  )
+
+  const announcements = {
+    onDragStart: id => {
+      const position = items.indexOf(id) + 1
+      const name = childrenDict[id].props.simpleName || id
+      return a11yAnnouncement.onDragStart(name, position, items.length)
     },
+    onDragOver: (id, overId) => {
+      const position = items.indexOf(overId) + 1
+      const name = childrenDict[id].props.simpleName || id
+      return a11yAnnouncement.onDragOver(name, position, items.length)
+    },
+    onDragEnd: (id, overId) => {
+      const position = items.indexOf(overId) + 1
+      const name = childrenDict[id].props.simpleName || id
+      return a11yAnnouncement.onDragEnd(name, position, items.length)
+    },
+    onDragCancel: () => a11yAnnouncement.cancel,
   }
 
-  useEffect(() => {
-    if (!listElement) return
-    const dragon = new DragonDrop(listElement.current, {
-      ...dragonOptions,
-      handle: showHandle && 'button.k-DragAndDropList__item__button',
-    })
-    dragon.on('dropped', handleChange)
+  const handleDragEnd = ({ active, over }) => {
+    setActiveId(null)
 
-    setDragonInstance(dragon)
-  }, [listElement, showHandle])
+    if (over && active.id !== over.id) {
+      const oldIndex = items.indexOf(active.id)
+      const newIndex = items.indexOf(over.id)
 
-  useEffect(() => {
-    if (!dragonInstance) return
-    if (!listElement) return
-    if (children.length == 0) return
-
-    dragonInstance.initElements(listElement.current)
-  }, [children, listElement, showHandle])
-
-  const handleChange = (container, item) => {
-    const newPosition = [...container.children].indexOf(item) + 1
-    const newList = [...container.children].map(child => {
-      return child.dataset?.id
-    })
-
-    onChange({ movedItem: item.dataset.id, newPosition, newList })
+      setItems(items => arrayMove(items, oldIndex, newIndex))
+      onChange({
+        movedItem: active.id,
+        newPosition: newIndex + 1,
+      })
+    }
   }
+
+  const handleDragStart = ({ active }) => {
+    if (!active) return
+    setActiveId(active.id)
+  }
+
+  const handleDragCancel = () => setActiveId(null)
 
   return (
     <>
-      <ol
-        ref={listElement}
-        className={classNames('k-DragAndDropList', className)}
-        aria-labelledby={a11yContainerLabelElement}
-        style={{
-          ...style,
-          '--dragAndDropList-gap': pxToRem(gap),
-        }}
-        {...props}
+      <DndContext
+        sensors={sensors}
+        collisonDetection={closestCenter}
+        onDragStart={handleDragStart}
+        onDragEnd={handleDragEnd}
+        onDragCancel={handleDragCancel}
+        modifiers={[restrictToVerticalAxis, restrictToWindowEdges]}
+        announcements={announcements}
+        screenReaderInstructions={a11yInstructions}
       >
-        {children?.length > 0 &&
-          [...children].map((child, index) => (
-            <li
-              key={child.props.id + index}
-              className={classNames('k-DragAndDropList__item', {
-                'k-DragAndDropList__item--hasButton': !!showHandle,
-              })}
-              data-simple-name={child.props.simpleName}
-              data-id={child.props.id}
-            >
-              {!!showHandle && (
-                <Button
-                  fit="content"
-                  aria-label={a11yButtonLabel}
-                  aria-describedby={a11yButtonDescElement}
-                  className="k-DragAndDropList__item__button"
-                >
-                  <MenuIcon width={10} height={10} />
-                </Button>
-              )}
-              {React.cloneElement(child, {
-                ...child.props,
-                className: classNames(
-                  child.props.className,
-                  'k-DragAndDropList__item__child',
-                ),
-              })}
-            </li>
-          ))}
-      </ol>
+        <ol
+          className={classNames('k-DragAndDropList', className)}
+          aria-labelledby={a11yContainerLabelElement}
+          style={{
+            ...style,
+            '--dragAndDropList-gap': pxToRem(gap),
+          }}
+          {...props}
+        >
+          <SortableContext items={items} strategy={verticalListSortingStrategy}>
+            {items.map(id => (
+              <SortableItem
+                key={id}
+                id={id}
+                showHandle={showHandle}
+                a11yButtonLabel={a11yButtonLabel}
+              >
+                {React.cloneElement(childrenDict[id])}
+              </SortableItem>
+            ))}
+          </SortableContext>
+        </ol>
+        {createPortal(
+          <DragOverlay>
+            {activeId ? (
+              <Item showHandle={showHandle} dragOverlay>
+                {React.cloneElement(childrenDict[activeId])}
+              </Item>
+            ) : null}
+          </DragOverlay>,
+          document.body,
+        )}
+      </DndContext>
       <DragAndDropListStyles />
     </>
   )
 }
 
+const SortableItem = ({ id, ...props }) => {
+  const {
+    attributes,
+    listeners,
+    setNodeRef,
+    transform,
+    transition,
+    isSorting,
+    isDragging,
+  } = useSortable({ id })
+
+  return (
+    <li
+      ref={setNodeRef}
+      id={id}
+      className={classNames('k-DragAndDropList__itemWrapper')}
+      style={{
+        transition,
+        '--translate-x': transform ? `${Math.round(transform.x)}px` : undefined,
+        '--translate-y': transform ? `${Math.round(transform.y)}px` : undefined,
+      }}
+    >
+      <Item
+        listeners={listeners}
+        attributes={attributes}
+        isSorting={isSorting}
+        isDragging={isDragging}
+        {...props}
+      />
+    </li>
+  )
+}
+
+const Item = ({
+  showHandle,
+  listeners,
+  attributes,
+  a11yButtonLabel,
+  children,
+  dragOverlay = false,
+  isSorting,
+  isDragging,
+}) => {
+  return (
+    <div
+      className={classNames('k-DragAndDropList__item', {
+        'k-DragAndDropList__item--hasButton': !!showHandle,
+        'k-DragAndDropList__item--dragOverlay': !!dragOverlay,
+        'k-DragAndDropList__item--isSorting': !!isSorting,
+        'k-DragAndDropList__item--isDragging': !!isDragging,
+      })}
+      {...(!showHandle ? listeners : undefined)}
+      tabIndex={!showHandle ? 0 : undefined}
+      {...attributes}
+    >
+      {!!showHandle && (
+        <Button
+          fit="content"
+          aria-label={a11yButtonLabel}
+          className="k-DragAndDropList__item__button"
+          modifier={!!dragOverlay ? 'helium' : 'hydrogen'}
+          {...listeners}
+        >
+          <MenuIcon width={10} height={10} />
+        </Button>
+      )}
+      <div className="k-DragAndDropList__item__child">{children}</div>
+    </div>
+  )
+}
+
 DragAndDropList.defaultProps = {
+  a11yInstructions: `To pick up a draggable item, press space or enter.
+    While dragging, use the arrow keys to move the item in any given direction.
+    Press space or enter again to drop the item in its new position,
+    or press escape to cancel.`,
   a11yButtonLabel: 'Reorder',
   a11yAnnouncement: {
-    grabbed: name => `${name} grabbed`,
-    dropped: name => `${name} dropped`,
-    reorder: (name, position, length) => {
-      return `The rankings have been updated, ${name} is now ranked #${position} out of ${length}`
+    onDragStart: (name, position, length) =>
+      `${name} grabbed from position ${position} of ${length}`,
+    onDragOver: (name, position, length) => {
+      return `${name} was moved to position ${position} of ${length}`
+    },
+    onDragEnd: (name, position, length) => {
+      return `${name} was dropped at position ${position} of ${length}`
     },
     cancel: 'Reranking cancelled.',
   },
@@ -196,11 +323,12 @@ DragAndDropList.defaultProps = {
 }
 
 DragAndDropList.propTypes = {
+  a11yInstructions: PropTypes.string,
   a11yButtonLabel: PropTypes.string,
   a11yAnnouncement: PropTypes.shape({
-    grabbed: PropTypes.func,
-    dropped: PropTypes.func,
-    reorder: PropTypes.func,
+    onDragStart: PropTypes.func,
+    onDragOver: PropTypes.func,
+    onDragEnd: PropTypes.func,
     cancel: PropTypes.string,
   }),
   onChange: PropTypes.func,