+++ CHANGELOG.md
@@ -4,6 +4,10 @@ This project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [unreleased]
 
+- Feature: `DragAndDropList`: Change lib.
+- Feature: `DragAndDropList`: Change `a11yAnnouncement` prop format.
+- Feature: `DragAndDropList`: Add `a11yInstructions` prop.
+
 ## [6.1.0] - 2021-11-15
 
 Features: